+++ tests/test.c
@@ -315,6 +315,8 @@ _TEST_DECL(0185_share_consumer_max_poll_interval);
 _TEST_DECL(0186_share_consumer_fatal_error);
 _TEST_DECL(0187_legacy_msgset_fetch_mock);
 _TEST_DECL(0190_share_consumer_telemetry);
+_TEST_DECL(0191_flexver_group_protocol_mock);
+_TEST_DECL(0192_sasl_authenticate_flexver);
 
 /* Manual tests */
 _TEST_DECL(8000_idle);
@@ -614,6 +616,8 @@ struct test tests[] = {
     _TEST(0190_share_consumer_telemetry,
           TEST_F_MANUAL,
           TEST_BRKVER(4, 2, 0, 0)),
+    _TEST(0191_flexver_group_protocol_mock, TEST_F_LOCAL),
+    _TEST(0192_sasl_authenticate_flexver, 0),
 
     /* Manual tests */
     _TEST(8000_idle, TEST_F_MANUAL),